APP_MEME/backend/requirements.txt
fastapi
uvicorn
requests
huggingface_hub
python-multipart
pillow
